+++ crates/oxide-update-engine-types/src/macros.rs
@@ -56,5 +56,7 @@ macro_rules! define_update_engine_types {
             $crate::events::EventReport<S>;
         $v type EventBuffer<S = $spec_type> =
             $crate::buffer::EventBuffer<S>;
+        $v type EventBufferStepData<S = $spec_type> =
+            $crate::buffer::EventBufferStepData<S>;
     };
 }